perf: Windows构建机的batch脚本执行编码问题 #4601
