+++ src/backend/ci/core/common/common-api/src/main/kotlin/com/tencent/devops/common/api/util/script/CommandLineUtils.kt
@@ -58,13 +58,6 @@ object CommandLineUtils {
         }
 
         val outputStream = object : LogOutputStream() {
-            override fun processBuffer() {
-                val privateStringField = LogOutputStream::class.java.getDeclaredField("buffer")
-                privateStringField.isAccessible = true;
-                val buffer = privateStringField.get(this) as ByteArrayOutputStream
-                processLine(String(buffer.toByteArray(), charset("UTF-8")))
-                buffer.reset()
-            }
             override fun processLine(line: String?, level: Int) {
                 if (line == null)
                     return
@@ -78,14 +71,6 @@ object CommandLineUtils {
         }
 
         val errorStream = object : LogOutputStream() {
-            override fun processBuffer() {
-                val privateStringField = LogOutputStream::class.java.getDeclaredField("buffer")
-                privateStringField.isAccessible = true;
-                val buffer = privateStringField.get(this) as ByteArrayOutputStream
-                processLine(String(buffer.toByteArray(), charset("UTF-8")))
-                buffer.reset()
-            }
-
             override fun processLine(line: String?, level: Int) {
                 if (line == null) {
                     return
