+++ src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/worker/common/utils/CommandLineUtils.kt
@@ -38,6 +38,7 @@ import org.apache.commons.exec.CommandLine
 import org.apache.commons.exec.LogOutputStream
 import org.apache.commons.exec.PumpStreamHandler
 import org.slf4j.LoggerFactory
+import java.io.ByteArrayOutputStream
 import java.io.File
 import java.util.regex.Pattern
 
@@ -67,6 +68,13 @@ object CommandLineUtils {
         val contextLogFile = if (!buildId.isNullOrBlank()) { ScriptEnvUtils.getContextFile(buildId) } else { null }
 
         val outputStream = object : LogOutputStream() {
+            override fun processBuffer() {
+                val privateStringField = LogOutputStream::class.java.getDeclaredField("buffer")
+                privateStringField.isAccessible = true;
+                val buffer = privateStringField.get(this) as ByteArrayOutputStream
+                processLine(String(buffer.toByteArray(), charset("GBK")))
+                buffer.reset()
+            }
             override fun processLine(line: String?, level: Int) {
                 if (line == null)
                     return
@@ -86,6 +94,13 @@ object CommandLineUtils {
         }
 
         val errorStream = object : LogOutputStream() {
+            override fun processBuffer() {
+                val privateStringField = LogOutputStream::class.java.getDeclaredField("buffer")
+                privateStringField.isAccessible = true;
+                val buffer = privateStringField.get(this) as ByteArrayOutputStream
+                processLine(String(buffer.toByteArray(), charset("GBK")))
+                buffer.reset()
+            }
             override fun processLine(line: String?, level: Int) {
                 if (line == null) {
                     return
